+++ doomed_patient/doomed_patient/Directions.h
@@ -1,10 +1,17 @@
+//! Class for representing compass directions.
+/*!
+  This class stores an enum for representing the four
+  compass directions, as well as static methods for
+  manipulating and getting information about the four 
+  directions.
+*/
+
 #pragma once
-#include "GridCoordinate.h"
+#include "VectorXY.h"
 
 class Directions
 {
 public:
-
 	//! Enum class for the four directions.
 	/*!
 	  This enum class is used to represent the
@@ -21,7 +28,18 @@ public:
 	Directions();
 	~Directions();
 
-	static GridCoordinate getDirectionVector(Direction direction);
+	//! Return the vector representative of the given direction
+	/*!
+	  This method returns the unit vector representative of the
+	  given Direction as a GridCoordinate.
+	*/
+	static VectorXY getDirectionVector(Direction direction);
+
+	//! Return the opposite direction to the given direction
+	/*!
+	  This method returns the direction that is opposite to
+	  the given direction as a Direction.
+	*/
 	static Direction getOpposite(Direction direction);
 };