+++ doomed_patient/doomed_patient/GameObject.cpp
@@ -6,8 +6,6 @@
 GameObject::GameObject(PatientGame* game, Texture* sprite)
 	: game(game), objectSprite(sprite)
 {
-	spriteSizeX = objectSprite->getSpriteSizeX();
-	spriteSizeY = objectSprite->getSpriteSizeY();
 }
 
 GameObject::~GameObject()
@@ -19,6 +17,6 @@ void GameObject::render(SDL_Renderer* renderer)
 	if (objectSprite)
 	{
 		// Render the game object
-		objectSprite->render(renderer, centreX, centreY, spriteSizeX, spriteSizeY);
+		objectSprite->render(renderer, centre);
 	}
 }