+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+##17/03/2016
+* Player now inherits from Character class
+* Player starting grid position is now defined by Player
+
+##16/03/2016
+* Created VectorXY class
+* All classes now use VectorXY for position instead of separate x and y integers
+
+##15/03/2016
+* Guards and Doctors now inherit from new Character class (which inherits from GameObject)
+* Guards and Doctors now generated using template function
+
+##09/03/2016
+* Exit is now generated on the top or right edge
+* All classes now use smart pointers where appropriate
+* Code in all classes has been tidied up and documented
+
 ##07/03/2016
 * Code in Level class has now been tidied up and documented